Arithmetic in VM

+++ CLox/chunk.c
@@ -24,7 +24,7 @@ void writeChunk(Chunk* chunk, uint8_t byte, int line) {
         chunk->capacity = GROW_CAPACITY(oldCapacity);
         chunk->code = GROW_ARRAY(uint8_t, 
             chunk->code, oldCapacity, chunk->capacity);
-        chunk->lines = GROW_ARRAY(uint8_t, 
+        chunk->lines = GROW_ARRAY(int, 
             chunk->lines, oldCapacity, chunk->capacity);
     }
